Extract DoiForm to own component, extract parseDoiMetaData to helper function
